publisher/requirements.txt
kafka-python==2.0.2
numpy==1.22.3
pip==23.0.1
setuptools==67.8.0
wheel==0.38.4
